Bug fixes and expanded attacker logic.
Fixes the implementation of tracking attacker/defender by row.  Bugs
were present because I didn't properly refresh my browser after the
last set of code changes.

Changes attacker logic to end the game if it gets to the left side
of the screen, attack if it gets next to a defender, or move left.

Noteably does not pause between attacks or queue up with other
attackers.